+++ packages/subtree-updater/src/sync/subgraph/adapter.ts
@@ -6,14 +6,14 @@ import {
   IncludedEncryptedNote,
   NoteTrait,
   TreeConstants,
-  IncludedNoteCommitment,
   TotalEntityIndex,
   maxArray,
   SubgraphUtils,
   TotalEntityIndexTrait,
   max,
+  WithTotalEntityIndex,
 } from "@nocturne-xyz/sdk";
-import { SubtreeUpdaterSyncAdapter } from "../syncAdapter";
+import { Insertion, SubtreeUpdaterSyncAdapter } from "../syncAdapter";
 import { fetchTreeInsertions, fetchLatestSubtreeIndex } from "./fetch";
 import { Logger } from "winston";
 
@@ -33,7 +33,7 @@ export class SubgraphSubtreeUpdaterSyncAdapter
   iterInsertions(
     startTotalEntityIndex: TotalEntityIndex,
     opts?: IterSyncOpts
-  ): ClosableAsyncIterator<IncludedNote | IncludedNoteCommitment> {
+  ): ClosableAsyncIterator<WithTotalEntityIndex<Insertion>> {
     const endpoint = this.graphqlEndpoint;
     const logger = this.logger;
     const endTotalEntityIndex = opts?.endTotalEntityIndex;
@@ -72,7 +72,7 @@ export class SubgraphSubtreeUpdaterSyncAdapter
         if (insertions.length > 0) {
           from = maxArray(sorted.map((i) => i.totalEntityIndex)) + 1n;
 
-          for (const { inner: insertion } of sorted) {
+          for (const { inner: insertion, totalEntityIndex } of sorted) {
             if ("numZeros" in insertion) {
               const startIndex = insertion.merkleIndex;
               const meta = {
@@ -86,8 +86,12 @@ export class SubgraphSubtreeUpdaterSyncAdapter
                 });
               for (let i = 0; i < insertion.numZeros; i++) {
                 yield {
-                  noteCommitment: TreeConstants.ZERO_VALUE,
-                  merkleIndex: startIndex + i,
+                  inner: {
+                    noteCommitment: TreeConstants.ZERO_VALUE,
+                    merkleIndex: startIndex + i,
+                  },
+                  // HACK: add `i` to `totalEntityIndex` to ensure all of the zeros have unique `totalEntityIndex`s
+                  totalEntityIndex: totalEntityIndex + BigInt(i),
                 };
               }
             } else if (NoteTrait.isEncryptedNote(insertion)) {
@@ -103,9 +107,13 @@ export class SubgraphSubtreeUpdaterSyncAdapter
                   insertionKind: "encrypted note",
                   meta,
                 });
+
               yield {
-                noteCommitment,
-                merkleIndex: insertion.merkleIndex,
+                inner: {
+                  noteCommitment,
+                  merkleIndex: insertion.merkleIndex,
+                },
+                totalEntityIndex,
               };
             } else {
               const meta = {
@@ -118,7 +126,10 @@ export class SubgraphSubtreeUpdaterSyncAdapter
                   meta,
                 });
 
-              yield insertion as IncludedNote;
+              yield {
+                inner: insertion as IncludedNote,
+                totalEntityIndex,
+              };
             }
           }
         } else {