+++ packages/subtree-updater/src/sync/syncAdapter.ts
@@ -4,14 +4,17 @@ import {
   IncludedNoteCommitment,
   IterSyncOpts,
   TotalEntityIndex,
+  WithTotalEntityIndex,
 } from "@nocturne-xyz/sdk";
+
+export type Insertion = IncludedNote | IncludedNoteCommitment;
 
 export interface SubtreeUpdaterSyncAdapter {
   // iterate over insertions, pulling them by merkle index
   iterInsertions(
     startTotalEntityIndex: TotalEntityIndex,
     opts?: IterSyncOpts
-  ): ClosableAsyncIterator<IncludedNote | IncludedNoteCommitment>;
+  ): ClosableAsyncIterator<WithTotalEntityIndex<Insertion>>;
 
   fetchLatestSubtreeIndex(): Promise<number | undefined>;
 }